pyproject.toml
[tool.poetry]
name = "dovado-rtl"
version = "0.9.11"
description = "RTL Design Space Exploration on top of Vivado"
authors = ["Daniele Paletti <danielepaletti98@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/DPaletti/dovado"
repository = "https://github.com/DPaletti/dovado"
keywords = [
    "Xilinx",
    "Vivado",
    "TCL",
    "VHDL",
    "Verilog",
    "RTL",
    "Design Space",
    "Genetic Algorithm",
]
packages = [
    { include = "dovado_rtl"},
    { include = "vhdl" },
    { include = "verilog" },
    { include = "xdc" },
    { include = "tcl" },
]
[tool.poetry.dependencies]
python = "^3.9"
pathvalidate = "^2.3.0"
BeautifulSoup4 = "^4.9.1"
lxml = "^4.5.2"
pyyaml = "^5.3.1"
pexpect = "^4.8.0"
pymoo = "^0.5.0"
antlr4-python3-runtime = "4.11.1"
dataclasses = { version = "^0.8", python = "~3.6" }
typer = "^0.3.2"
#movado = { path = "../movado/dist/Movado-0.6.3-py3-none-any.whl" }
importlib-resources = "^5.1.4"
nacolla = "*"
movado = "0.7.4"
numpy = "1.23.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"

[tool.poetry.scripts]
dovado = "dovado_rtl.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
addopts = "--ignore=tests/resources --rootdir=tests"

[tool.pyright]
exclude = ["tests/resources"]
